+++ v8/protocol/src/org/chromium/sdk/internal/v8native/protocol/input/data/ScriptWithId.java
@@ -4,5 +4,5 @@ import org.chromium.protocolReader.JsonType;
 
 @JsonType
 public interface ScriptWithId {
-  long id();
+  int id();
 }
